@@ -11,17 +11,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.1.3",
+    "@angular/animations": "^7.1.3",
+    "@angular/cdk": "^7.3.7",
     "@angular/common": "7.0.0",
     "@angular/compiler": "7.0.0",
-    "@angular/core": "7.0.0",
+    "@angular/core": "^7.2.0",
+    "@angular/flex-layout": "^7.0.0-beta.24",
     "@angular/forms": "7.0.0",
     "@angular/http": "7.1.3",
+    "@angular/material": "^7.3.7",
     "@angular/platform-browser": "7.0.0",
     "@angular/platform-browser-dynamic": "7.0.0",
     "@angular/router": "7.0.0",
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^4.0.0",
     "@types/jquery": "^3.3.29",
     "core-js": "2.5.7",
+    "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "jqwidgets-scripts": "6.2.0",
     "materialize-css": "^1.0.0-rc.2",
