@@ -2,6 +2,7 @@ import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {SharedModule} from './shared/shared.module';
 import {SiteModule} from './shared/layouts/site-layout/site.module';
 
@@ -12,6 +13,8 @@ import {AuthLayoutComponent} from './shared/layouts/auth-layout/auth-layout.comp
 import {RegisterPageComponent} from './auth/register-page/register-page.component';
 import {TokenInterceptor} from './shared/classes/token.interceptor';
 import {NotFoundComponent} from './shared/components/not-found/not-found.component';
+import {MaterialModule} from './shared/material-module';
+
 
 @NgModule({
   declarations: [
@@ -20,16 +23,20 @@ import {NotFoundComponent} from './shared/components/not-found/not-found.compone
     AuthLayoutComponent,
     RegisterPageComponent,
     NotFoundComponent,
+
+
   ],
   imports: [
     BrowserModule,
     FormsModule,
     ReactiveFormsModule,
     HttpClientModule,
+    BrowserAnimationsModule,
 
     AppRoutingModule,
     SharedModule,
-    SiteModule
+    MaterialModule,
+    SiteModule,
   ],
   providers: [
     {
@@ -42,3 +49,4 @@ import {NotFoundComponent} from './shared/components/not-found/not-found.compone
 })
 export class AppModule {
 }
+
