@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient, HttpHeaders, HttpParams} from '@angular/common/http';
 import {Observable} from 'rxjs/index';
 
-import {NodeType, Message} from '../../interfaces';
+import {Contract, Message} from '../../interfaces';
 
 @Injectable({
   providedIn: 'root'
@@ -15,8 +15,8 @@ export class NodeTypeService {
 
   // get
 
-  getAll(params: any = {}): Observable<NodeType[]> {
-    return this.http.get<NodeType[]>('/api/v1/nodes-types', {
+  getAll(params: any = {}): Observable<Contract[]> {
+    return this.http.get<Contract[]>('/api/v1/nodes-types', {
       params: new HttpParams({
         fromObject: params
       })
@@ -25,17 +25,14 @@ export class NodeTypeService {
 
   // post
 
-  ins(nodeType: NodeType): Observable<NodeType> {
-    return this.http.post<NodeType>('/api/v1/nodes-types', nodeType);
+  ins(nodeType: Contract): Observable<Contract> {
+    return this.http.post<Contract>('/api/v1/nodes-types', nodeType);
   }
 
   // patch
 
-  upd(nodeType: NodeType): Observable<NodeType> {
-
-    console.log(nodeType)
-
-    return this.http.patch<NodeType>('/api/v1/nodes-types', nodeType);
+  upd(nodeType: Contract): Observable<Contract> {
+    return this.http.patch<Contract>('/api/v1/nodes-types', nodeType);
   }
 
   // delete
