@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient, HttpParams} from '@angular/common/http';
 import {Observable} from 'rxjs/index';
 
-import {EquipmentType, Message, NodeType, SensorType} from '../../interfaces';
+import {EquipmentType, Message, Contract, SensorType} from '../../interfaces';
 
 @Injectable({
   providedIn: 'root'
@@ -28,21 +28,18 @@ export class SensorTypeService {
   // post
 
   ins(sensorType: SensorType): Observable<SensorType> {
-    // return this.http.post<EquipmentType>('/api/v1/nodes', nodeType);
-    return null;
+    return this.http.post<SensorType>('/api/v1/sensors-types', sensorType);
   }
 
   // patch
 
   upd(sensorType: SensorType): Observable<SensorType> {
-    // return this.http.patch<EquipmentType>('/api/v1/nodes', nodeType);
-    return null;
+    return this.http.patch<SensorType>('/api/v1/sensors-types', sensorType);
   }
 
   // delete
 
   del(id_sensorType: number): Observable<Message> {
-    // return this.http.delete<Message>(`/api/v1/nodes/${id_nodeType}`);
-    return null;
+    return this.http.delete<Message>(`/api/v1/sensors-types/${id_sensorType}`);
   }
 }
